refactor(upgrade): extract managed-pair engine
