+++ BUILD.bazel
@@ -124,6 +124,7 @@ CARGO_WORKSPACE_PACKAGE_DATA = [
     "//crates/ctx-history-source-discovery:cargo_package_data",
     "//crates/ctx-history-source-io:cargo_package_data",
     "//crates/ctx-history-source-sqlite:cargo_package_data",
+    "//crates/ctx-managed-pair-engine:cargo_package_data",
     "//crates/ctx-semantic-index:cargo_package_data",
     "//crates/ctx-semantic-model:cargo_package_data",
     "//crates/ctx-terminal:cargo_package_data",
@@ -1085,6 +1086,8 @@ sh_test(
         "//crates/ctx-cli:cargo_package_data",
         "//crates/ctx-history-core:BUILD.bazel",
         "//crates/ctx-history-core:cargo_package_data",
+        "//crates/ctx-managed-pair-engine:BUILD.bazel",
+        "//crates/ctx-managed-pair-engine:cargo_package_data",
         "//crates/ctx-upgrade-engine:BUILD.bazel",
         "//crates/ctx-upgrade-engine:cargo_package_data",
     ],
