+++ crates/ctx-managed-pair-engine/BUILD.bazel
@@ -1,0 +1,54 @@
+load("@crates//:defs.bzl", "aliases", "all_crate_deps", "crate_edition")
+load("@rules_rust//rust:defs.bzl", "rust_library")
+load("//:rust_sources.bzl", "RUST_PROD_SRC_EXCLUDES")
+load("//tools/bazel:ctx_rust.bzl", "ctx_rust_test")
+load("//tools/bazel:gates.bzl", "loc_check_inputs")
+
+package(default_visibility = ["//visibility:public"])
+
+loc_check_inputs(name = "loc_check_inputs")
+
+filegroup(
+    name = "cargo_package_data",
+    srcs = glob(["**"], exclude = ["BUILD.bazel"]),
+)
+
+COMMON_RUSTC_ENV = {
+    "CARGO_MANIFEST_DIR": "crates/ctx-managed-pair-engine",
+}
+
+RUST_SRCS = glob(["src/**/*.rs"])
+
+PROD_SRCS = glob(
+    ["src/**/*.rs"],
+    exclude = RUST_PROD_SRC_EXCLUDES,
+)
+
+CTX_MANAGED_PAIR_ENGINE_DEPS = [
+    "//crates/ctx-history-platform:lib",
+]
+
+rust_library(
+    name = "lib",
+    srcs = PROD_SRCS,
+    crate_name = "ctx_managed_pair_engine",
+    crate_root = "src/lib.rs",
+    edition = crate_edition(),
+    aliases = aliases(),
+    deps = all_crate_deps(normal = True) + CTX_MANAGED_PAIR_ENGINE_DEPS,
+    proc_macro_deps = all_crate_deps(proc_macro = True),
+    rustc_env = COMMON_RUSTC_ENV,
+)
+
+ctx_rust_test(
+    name = "unit_tests",
+    srcs = RUST_SRCS,
+    crate_name = "ctx_managed_pair_engine",
+    crate_root = "src/lib.rs",
+    edition = crate_edition(),
+    aliases = aliases(normal_dev = True, proc_macro_dev = True),
+    deps = all_crate_deps(normal = True, normal_dev = True) + CTX_MANAGED_PAIR_ENGINE_DEPS,
+    proc_macro_deps = all_crate_deps(proc_macro = True, proc_macro_dev = True),
+    rustc_env = COMMON_RUSTC_ENV,
+    tags = ["exclusive"],
+)
