+++ crates/ctx-managed-pair-engine/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "ctx-upgrade-engine"
+name = "ctx-managed-pair-engine"
 version.workspace = true
-description = "Internal release upgrade engine for ctx"
+description = "Internal verifier-driven managed Core and companion publication engine for ctx"
 edition.workspace = true
 license.workspace = true
 repository.workspace = true
@@ -9,30 +9,16 @@ homepage.workspace = true
 rust-version.workspace = true
 publish = false
 
-[features]
-test-support = []
-
 [dependencies]
 anyhow.workspace = true
-base64.workspace = true
-chrono.workspace = true
-ctx-history-core = { path = "../ctx-history-core" }
 ctx-history-platform = { path = "../ctx-history-platform" }
-ctx-managed-pair-engine = { path = "../ctx-managed-pair-engine" }
 fs2 = "0.4.3"
-ring.workspace = true
-semver.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
-tar.workspace = true
-url.workspace = true
 uuid.workspace = true
-xz2.workspace = true
-zstd.workspace = true
 
 [target.'cfg(unix)'.dependencies]
-flate2 = "1"
 libc.workspace = true
 
 [target.'cfg(windows)'.dependencies]
