+++ crates/ctx-managed-pair-engine/src/attempt.rs
