+++ crates/ctx-managed-pair-engine/src/journal.rs
