+++ crates/ctx-managed-pair-engine/src/uninstall.rs
