+++ crates/ctx-terminal/tests/raw_output_policy/owning_test.rs
@@ -65,6 +65,7 @@ fn parse_identity(identity: &str) -> Result<(String, String), String> {
         && !path.starts_with("crates/ctx-agent-integrations/src/")
         && !path.starts_with("crates/ctx-client-observability/src/")
         && !path.starts_with("crates/ctx-history-cli/src/")
+        && !path.starts_with("crates/ctx-managed-pair-engine/src/")
         && !path.starts_with("crates/ctx-upgrade-engine/src/")
     {
         return Err("owning test source is outside the source-checked test roots".to_owned());
