+++ tools/bazel/test_tiers.bzl
@@ -92,6 +92,7 @@ RUST_FORMAT_TARGETS = [
     "//crates/ctx-history-providers-task-docs:unit_tests",
     "//crates/ctx-history-refresh:unit_tests",
     "//crates/ctx-history-refresh-execution:unit_tests",
+    "//crates/ctx-managed-pair-engine:unit_tests",
     "//crates/ctx-semantic-index:unit_tests",
     "//crates/ctx-semantic-model:unit_tests",
     "//crates/ctx-terminal:unit_tests",
@@ -222,6 +223,7 @@ _CI_RUST_TESTS = [
     "//tools/bazel:history_provider_pack_boundary_mutation_tests",
     "//crates/ctx-history-refresh:unit_tests",
     "//crates/ctx-history-refresh-execution:unit_tests",
+    "//crates/ctx-managed-pair-engine:unit_tests",
     "//crates/ctx-semantic-index:unit_tests",
     "//crates/ctx-semantic-model:unit_tests",
     "//crates/ctx-terminal:unit_tests",
